# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Captures By Capri — Photography portfolio and booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GalleryImage {
  id: number;
  title: string;
  url: string;
  category: string;
  displayOrder: number;
  featured?: boolean;
  createdAt: string;
}

export interface GalleryImageInput {
  title: string;
  url: string;
  category: string;
  displayOrder?: number;
  featured?: boolean;
}

export interface GalleryCategory {
  name: string;
  count: number;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  approved: 'approved',
  declined: 'declined',
} as const;

export interface Booking {
  id: number;
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  sessionType: string;
  preferredDate: string;
  /** @nullable */
  preferredTime?: string | null;
  /** @nullable */
  message?: string | null;
  status: BookingStatus;
  /** @nullable */
  calBookingId?: string | null;
  /** @nullable */
  declineReason?: string | null;
  smsSent?: boolean;
  adminNotified?: boolean;
  createdAt: string;
}

export interface BookingInput {
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  sessionType: string;
  preferredDate: string;
  preferredTime?: string;
  message?: string;
  calBookingId?: string;
}

export interface DeclineInput {
  reason?: string;
}

export interface BookingStats {
  total: number;
  pending: number;
  approved: number;
  declined: number;
  thisMonth: number;
}

export type CalcomWebhookPayloadPayloadAttendeesItem = {
  name?: string;
  email?: string;
  timeZone?: string;
};

export type CalcomWebhookPayloadPayloadResponses = { [key: string]: unknown };

export type CalcomWebhookPayloadPayload = {
  uid?: string;
  title?: string;
  startTime?: string;
  endTime?: string;
  attendees?: CalcomWebhookPayloadPayloadAttendeesItem[];
  responses?: CalcomWebhookPayloadPayloadResponses;
};

export interface CalcomWebhookPayload {
  triggerEvent: string;
  payload: CalcomWebhookPayloadPayload;
}

export type ListGalleryParams = {
category?: string;
};

export type ListBookingsParams = {
status?: ListBookingsStatus;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  pending: 'pending',
  approved: 'approved',
  declined: 'declined',
} as const;