# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Captures By Capri — Photography portfolio and booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  sessionType: string;
  preferredDate: string;
  /** @nullable */
  preferredTime?: string | null;
  /** @nullable */
  message?: string | null;
  status: BookingStatus;
  /** @nullable */
  calBookingId?: string | null;
  /** @nullable */
  declineReason?: string | null;
  smsSent?: boolean;
  adminNotified?: boolean;
  createdAt: string;
}