# lib/api-zod/src/generated/types/declineInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Captures By Capri — Photography portfolio and booking API
 * OpenAPI spec version: 0.1.0
 */

export interface DeclineInput {
  reason?: string;
}