# lib/api-zod/src/generated/types/galleryImageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Captures By Capri — Photography portfolio and booking API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryImageInput {
  title: string;
  url: string;
  category: string;
  displayOrder?: number;
  featured?: boolean;
}